Trying to cleanup the amount of code in utils in the backend. Split utils into two files (commonutils and utils). docsrender (backend) only depends on common utils.
